requirements.txt
discord.py==2.1.1
Flask==2.1.0
Werkzeug==2.0.3
supabase
matplotlib==3.5.1
